langchain-module/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
openai
python-dotenv
requests
PyGithub
python-gitlab
bitbucket-python
pydantic
click
rich
tiktoken 